+++ twodee/matt/utils/solve.go
